requirements.txt
torch
transformers
music21
pretty_midi
soundfile
Pillow
textblob
midi2audio
